requirements.txt
cx_Oracle==8.3.0
reportlab==4.0.8
python-dotenv==1.0.0
